requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & NLP
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0

# LLM API
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipython>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

